chore: cleanup event loop utils
