+++ src/common/utils/Messenger.ts
@@ -1,6 +1,6 @@
 import browser, { Runtime } from 'webextension-polyfill';
-import { sleep } from './EventLoop';
 import { debug } from './log';
+import { sleep } from './time';
 
 export default class Messenger<
   K extends MessageTypes = MessageTypes,
