+++ src/content-scripts/player/components/Player.vue
@@ -39,7 +39,7 @@
 <script lang="ts" setup>
 import { useTimeout } from '@anime-skip/ui';
 import { PLAYER_ACTIVITY_TIMEOUT } from '~/common/utils/Constants';
-import { nextFrame } from '~/common/utils/EventLoop';
+import { nextFrame } from '~/common/utils/event-loop';
 import Messenger from '~/common/utils/Messenger';
 import UsageStats from '~/common/utils/UsageStats';
 import Utils from '~/common/utils/Utils';
