+++ src/content-scripts/services/funimation-2021-09-26/parent.ts
@@ -1,6 +1,6 @@
-import { waitUntil } from '~/common/utils/EventLoop';
 import { loadedLog, log } from '~/common/utils/log';
 import setupParent from '~/common/utils/setupParent';
+import { waitUntil } from '~/common/utils/time';
 
 async function getEpisodeInfo() {
   const pageHasLoaded = () => Promise.resolve(document.querySelector('.meta-overlay') != null);
