+++ src/popup/composition/useLogout.ts
@@ -1,6 +1,6 @@
 import { useClearTokens } from '~/common/state/useAuth';
 import { useResetPreferences } from '~/common/state/useGeneralPreferences';
-import { sleep } from '~/common/utils/EventLoop';
+import { sleep } from '~/common/utils/time';
 import UsageStats from '~/common/utils/UsageStats';
 
 export function useLogout() {
